+++ examples/java/DonutIO/StartStage.java
@@ -37,7 +37,7 @@ public class StartStage extends Stage {
   public void whenKeyPressed(int keyCode) {
     if (keyCode == KeyCode.VK_SPACE) {
       Game.LEVEL = 0;
-      Window.getInstance().setStage(new WorldStage());
+      Window.getInstance().transitionToStage(new WorldStage(), 500);
     }
   }
 }